Revert "refactor fleet mode detection and storage (#40667)"
This reverts commit 33096202ff90c79daeb59447e61ff3430a6b0ea2.
