+++ libbeat/common/fleetmode/fleet_mode.go
@@ -17,18 +17,33 @@
 
 package fleetmode
 
-var managementEnabled bool
+import (
+	"flag"
 
-// SetAgentMode stores if the Beat is running under Elastic Agent.
-// Normally this is called when the command line flags are parsed.
-// This is stored as a package level variable because some components
-// (like filebeat/metricbeat modules) don't have access to the
-// configuration information to determine this on their own.
-func SetAgentMode(enabled bool) {
-	managementEnabled = enabled
-}
+	"github.com/elastic/elastic-agent-libs/config"
+)
 
-// Enabled returns true if the Beat is running under Elastic Agent.
+// Enabled checks to see if filebeat/metricbeat is running under Agent
+// The management setting is stored in the main Beat runtime object, but we can't see that from a module
+// So instead we check the CLI flags, since Agent starts filebeat/metricbeat with "-E", "management.enabled=true"
 func Enabled() bool {
-	return managementEnabled
+	type management struct {
+		Enabled bool `config:"management.enabled"`
+	}
+	var managementSettings management
+
+	cfgFlag := flag.Lookup("E")
+	if cfgFlag == nil {
+		return false
+	}
+
+	cfgObject, _ := cfgFlag.Value.(*config.SettingsFlag)
+	cliCfg := cfgObject.Config()
+
+	err := cliCfg.Unpack(&managementSettings)
+	if err != nil {
+		return false
+	}
+
+	return managementSettings.Enabled
 }
